Add fluids and vasopressin to feature vector
Validate that the time series labels is of compatible size with
the data. If the difference is small enough, it is corrected and
returned. If the difference is too large, the hadm_id is dropped.

Drop the *first* label time step and the *last* feature time step.
This is because labels will correspond to the feature data from
the previous time step. So, in the first time step, the label is
from the second time step, and in the last time step, we have no
label, so we drop that step. This also addresses the issue of
dosages going to zero if the patient dies.